Use docker to install prisma layer deps
