+++ src/cdk/lambda/prismaLayer/Dockerfile
@@ -1,0 +1,6 @@
+FROM amazon/aws-lambda-nodejs:14
+
+COPY package.json /asset/nodejs/
+WORKDIR /asset
+RUN cd nodejs && npm install
+
