+++ src/cdk/lambda/prismaLayer/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "prismalayer",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@prisma/client": "^3.1.1",
+    "@prisma/migrate": "^3.1.1",
+    "@prisma/sdk": "^3.1.1",
+    "prisma": "^3.1.1"
+  }
+}
