+++ src/cdk/lambda/prismaLayer/prismaLayer.test.ts
@@ -1,0 +1,20 @@
+import "@aws-cdk/assert/jest"
+import { Stack } from "@aws-cdk/core"
+import * as path from "path"
+import { PrismaLayer } from "./prismaLayer"
+
+const prismaPath = path.join(__dirname)
+
+describe("PrismaLayer", () => {
+  let stack: Stack
+
+  beforeEach(() => {
+    stack = new Stack()
+  })
+
+  it("builds layer", () => {
+    new PrismaLayer(stack, "PL", { prismaPath })
+
+    expect(stack).toHaveResource("AWS::Lambda::LayerVersion", {})
+  })
+})
