+++ src/cdk/lambda/prismaLayer/schema.prisma
@@ -1,0 +1,4 @@
+generator client {
+          provider      = "prisma-client-js"
+          binaryTargets = ["rhel-openssl-1.0.x"]
+}
